+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * Added support to WeNet OAuth2 authentication flow.
 * Integrated survey in the app.
 * Created a web service for parsing survey responses.
+* Set up the structure for updating a user profile.